backend/ocr_backend/requirements.txt
absl-py==2.3.0
aiobotocore==2.17.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-cors==0.8.1
aioitertools==0.12.0
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.6
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
beartype==0.15.0
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==4.2.1
biopython==1.84
bleach==6.2.0
blinker==1.9.0
bokeh==3.7.3
boto3==1.35.93
botocore==1.35.93
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.0
decorator==5.2.1
deltalake==0.17.4
deprecation==2.1.0
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
Django==5.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-redis==5.4.0
django-sslserver==0.22
django-treebeard==4.7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
easyocr==1.7.2
et_xmlfile==2.0.0
eventlet==0.39.1
executing==2.2.0
filelock==3.18.0
fire==0.7.0
firebase-admin==6.9.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
flower==2.0.1
fonttools==4.57.0
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.3.2
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-pubsub==2.30.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.12.0
greenlet==3.1.1
groq==0.25.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
h3==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.15
keras==3.10.0
kiwisolver==1.4.7
kombu==5.5.4
langchain==0.3.25
langchain-core==0.3.67
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
lmdb==1.6.2
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.4.3
mysql-connector-python==9.0.0
namex==0.1.0
narwhals==1.44.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opt-einsum==3.3.0
optree==0.16.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.0
paddleocr==2.10.0
paddlepaddle==3.0.0
pandas==2.3.0
panel==1.7.2
param==2.2.1
parso==0.8.4
pathway==0.24.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.2.2
pluggy==1.5.0
postgrest==1.0.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
PyMySQL==1.1.1
pyparsing==3.2.3
pytesseract==0.3.13
pytest==8.4.1
pytest-mock==3.14.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-engineio==4.11.2
python-sat==1.8.dev17
python-socketio==5.12.1
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.2
RapidFuzz==3.13.0
razorpay==1.4.2
realtime==2.4.3
redis==6.2.0
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.10.4
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.3.1
shapely==2.1.0
simple-websocket==1.1.0
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.39
sqlparse==0.5.3
stack-data==0.6.3
storage3==0.11.3
StrEnum==0.4.15
stringzilla==3.12.5
supabase==2.15.3
supafunc==0.9.4
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.4.0
uv==0.7.19
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
xyzservices==2025.4.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
